Revert "Refactored Profile page"
