--- frontend/src/app/profile/page.tsx
@@ -1,11 +1,0 @@
-
-import ProfileInfo from "@/components/profile/profileinfo"
-
-
-export default function ProfilePage() {
-  return (
-
-    <ProfileInfo />
-
-  )
-}
